FIX[backend]: fixed bpm ranges of dances in initDanceTypes() (Issue #23)

+++ backend/src/main/java/at/leonding/htl/features/dance/Dance.java
@@ -47,7 +47,7 @@ public class Dance {
         this.maxBpm = maxBpm;
     }
 
-    public boolean isBpmInRange(int bpmToCheck){
+    public boolean isBpmInRange(double bpmToCheck){
         return bpmToCheck >= minBpm && bpmToCheck <= maxBpm;
     }
